refactor: merge create and update acc into unic upsert component
