--- NERBABO.Frontend/src/app/features/acc/update-acc/update-acc.component.html
@@ -1,177 +1,0 @@
-<!-- button that triggers modal -->
-<div class="modal-header">
-  <h4 class="modal-title pull-left">Atualizar Utilizador</h4>
-  <button
-    type="button"
-    class="btn-close"
-    aria-label="Close"
-    (click)="bsModalRef.hide()"
-  ></button>
-</div>
-<div class="modal-body">
-  <!-- resgistration form -->
-  <form [formGroup]="updateUserForm">
-    <div class="property">
-      <strong>Nome da Pessoa</strong>
-      <p class="text-break text-muted mb-0">
-        {{ currentUser.fullName }}
-      </p>
-    </div>
-    <!-- email -->
-    <div class="mb-3">
-      <label for="email"
-        >Email
-        <span class="text-danger">*</span>
-      </label>
-      <input
-        type="text"
-        pInputText
-        formControlName="email"
-        [class.is-invalid]="submitted && updateUserForm.get('email')?.errors"
-        class="w-100 {{
-          submitted && updateUserForm.get('email')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
-        }}"
-      />
-
-      @if (submitted && updateUserForm.get('email')?.hasError("required")) {
-      <span class="text-danger"> Email é um campo obrigatório. </span>
-      } @else if (submitted &&
-      updateUserForm.get('email')?.hasError("minlength") ||
-      updateUserForm.get('email')?.hasError("maxlength")) {
-      <span class="text-danger"> Email deve ter entre 5 e 50 caracteres. </span>
-      } @else if (submitted && updateUserForm.get('email')?.hasError("email")) {
-      <span class="text-danger"> Email inválido. </span>
-      }
-    </div>
-    <!-- end email -->
-
-    <!-- username -->
-    <div class="mb-3">
-      <label for="userName">Username <span class="text-danger">*</span></label>
-      <input
-        type="text"
-        pInputText
-        formControlName="userName"
-        [class.is-invalid]="submitted && updateUserForm.get('username')?.errors"
-        class="w-100 {{
-          submitted && updateUserForm.get('userName')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
-        }}"
-      />
-      @if (submitted && updateUserForm.get('userName')?.hasError("required")) {
-      <span class="text-danger"> Username é um campo obrigatório. </span>
-      } @else if (submitted &&
-      updateUserForm.get('userName')?.hasError("minlength") ||
-      updateUserForm.get('userName')?.hasError("maxlength")) {
-      <span class="text-danger">
-        Username deve ter entre 5 e 50 caracteres.
-      </span>
-      }
-    </div>
-    <!-- end username -->
-
-    <!-- password -->
-    <div class="mb-3">
-      <label for="password">Password <span class="text-danger">*</span></label>
-      <p-password
-        formControlName="password"
-        [toggleMask]="true"
-        [class.is-invalid]="submitted && updateUserForm.get('password')?.errors"
-        class="{{
-          submitted && updateUserForm.get('password')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
-        }}"
-        styleClass="w-100"
-      >
-        <ng-template #header>
-          <div class="font-semibold text-xm mb-4">
-            Uma password fácil de relembrar
-          </div>
-        </ng-template>
-        <ng-template #footer>
-          <p-divider />
-          <ul class="pl-2 my-0 leading-normal">
-            <li>Mínimo de 8 caracteres;</li>
-            <li>Pelo menos 1 carácter especial (ex: !, ., #, $, %, etc.);</li>
-            <li>Incluir 1 letra maiúscula e 1 letra minúscula;</li>
-            <li>Incluir pelo menos 1 número.</li>
-          </ul>
-        </ng-template></p-password
-      >
-
-      @if (submitted && updateUserForm.get('password')?.hasError("required")) {
-      <span class="text-danger"> Password é um campo obrigatório. </span>
-      } @else if (submitted &&
-      updateUserForm.get('password')?.hasError("minlength") ||
-      updateUserForm.get('password')?.hasError("maxlength")) {
-      <span class="text-danger">
-        Password deve ter entre 5 e 50 caracteres.
-      </span>
-      }
-    </div>
-    <!-- end password -->
-
-    <!-- confirm password -->
-    <div class="mb-3">
-      <label for="confirmPassword"
-        >Confirmação Password <span class="text-danger">*</span></label
-      >
-      <p-password
-        autocomplete="off"
-        formControlName="confirmPassword"
-        [toggleMask]="true"
-        [class.is-invalid]="
-          submitted && updateUserForm.get('confirmPassword')?.errors
-        "
-        styleClass="w-100"
-        class=" {{
-          submitted && updateUserForm.get('confirmPassword')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
-        }}"
-      >
-        <ng-template #header>
-          <div class="font-semibold text-xm mb-4">Password de Confirmação</div>
-        </ng-template>
-        <ng-template #footer>
-          <p-divider />
-          <ul class="pl-2 my-0 leading-normal">
-            <li>Mínimo de 8 caracteres;</li>
-            <li>Pelo menos 1 carácter especial (ex: !, ., #, $, %, etc.);</li>
-            <li>Incluir 1 letra maiúscula e 1 letra minúscula;</li>
-            <li>Incluir pelo menos 1 número.</li>
-          </ul>
-        </ng-template></p-password
-      >
-      @if (submitted &&
-      updateUserForm.get('confirmPassword')?.hasError("required")) {
-      <span class="text-danger">
-        Confirm Password é um campo obrigatório.
-      </span>
-      } @if (submitted && updateUserForm.errors?.['mismatchedPasswords']) {
-      <span class="text-danger"> As palavras-passe não coincidem. </span>
-      }
-    </div>
-    <!-- end confirm password -->
-
-    <app-error-card [errors]="errorMessages" [submitted]="submitted" />
-  </form>
-</div>
-<div class="modal-footer">
-  <!-- submit button -->
-  <div class="text-end">
-    <button type="button" class="btn-gray me-2" (click)="bsModalRef.hide()">
-      Cancelar
-    </button>
-    <button type="submit" class="btn-main" (click)="onSubmit()">
-      @if (loading){
-      <span class="spinner-border spinner-border-sm me-1"></span>
-      } Registar
-    </button>
-  </div>
-  <!-- end submit button -->
-</div>
