+++ NERBABO.Frontend/src/app/features/acc/upsert-acc/upsert-acc.component.css
@@ -1,0 +1,7 @@
+:host ::ng-deep .p-password {
+  width: 100% !important;
+  max-width: 100% !important;
+}
+:host ::ng-deep .p-password input {
+  width: 100% !important;
+}
