+++ NERBABO.Frontend/src/app/features/acc/upsert-acc/upsert-acc.component.html
@@ -1,6 +1,8 @@
 <!-- button that triggers modal -->
 <div class="modal-header">
-  <h4 class="modal-title pull-left">Registar Utilizador</h4>
+  <h4 class="modal-title pull-left">
+    {{ isUpdate ? "Atualizar" : "Registar" }} Utilizador
+  </h4>
   <button
     type="button"
     class="btn-close"
@@ -10,10 +12,18 @@
 </div>
 <div class="modal-body">
   <!-- resgistration form -->
-  <form [formGroup]="registrationForm">
+  <form [formGroup]="form">
     <!-- person -->
+    @if(isUpdate){
+    <div class="property">
+      <strong>Nome da Pessoa</strong>
+      <p class="text-break text-muted mb-0">
+        {{ currentUser?.fullName }}
+      </p>
+    </div>
+    } @else {
     <div class="mb-3">
-      <label for="email"
+      <label for="personId"
         >Pessoa
         <span class="text-danger">*</span>
       </label>
@@ -24,21 +34,17 @@
         optionLabel="displayName"
         placeholder="Escolha uma pessoa..."
         [dropdown]="true"
-        [class.is-invalid]="
-          submitted && registrationForm.get('personId')?.errors
-        "
+        [class.is-invalid]="submitted && form.get('personId')?.errors"
         styleClass="w-100"
         class="{{
-          submitted && registrationForm.get('personId')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
+          submitted && form.get('personId')?.errors ? 'ng-invalid ng-dirty' : ''
         }}"
       />
-      @if (submitted && registrationForm.get('personId')?.hasError("required"))
-      {
+      @if (submitted && form.get('personId')?.hasError("required")) {
       <span class="text-danger"> Pessoa é um campo obrigatório. </span>
       }
     </div>
+    }
     <!-- end person -->
 
     <!-- email -->
@@ -51,22 +57,18 @@
         type="text"
         pInputText
         formControlName="email"
-        [class.is-invalid]="submitted && registrationForm.get('email')?.errors"
+        [class.is-invalid]="submitted && form.get('email')?.errors"
         class="w-100 {{
-          submitted && registrationForm.get('email')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
+          submitted && form.get('email')?.errors ? 'ng-invalid ng-dirty' : ''
         }}"
       />
 
-      @if (submitted && registrationForm.get('email')?.hasError("required")) {
+      @if (submitted && form.get('email')?.hasError("required")) {
       <span class="text-danger"> Email é um campo obrigatório. </span>
-      } @else if (submitted &&
-      registrationForm.get('email')?.hasError("minlength") ||
-      registrationForm.get('email')?.hasError("maxlength")) {
+      } @else if (submitted && form.get('email')?.hasError("minlength") ||
+      form.get('email')?.hasError("maxlength")) {
       <span class="text-danger"> Email deve ter entre 5 e 50 caracteres. </span>
-      } @else if (submitted && registrationForm.get('email')?.hasError("email"))
-      {
+      } @else if (submitted && form.get('email')?.hasError("email")) {
       <span class="text-danger"> Email inválido. </span>
       }
     </div>
@@ -74,26 +76,20 @@
 
     <!-- username -->
     <div class="mb-3">
-      <label for="username">Username <span class="text-danger">*</span></label>
+      <label for="userName">Username <span class="text-danger">*</span></label>
       <input
         type="text"
         pInputText
-        formControlName="username"
-        [class.is-invalid]="
-          submitted && registrationForm.get('username')?.errors
-        "
+        formControlName="userName"
+        [class.is-invalid]="submitted && form.get('userName')?.errors"
         class="w-100 {{
-          submitted && registrationForm.get('username')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
+          submitted && form.get('userName')?.errors ? 'ng-invalid ng-dirty' : ''
         }}"
       />
-      @if (submitted && registrationForm.get('userName')?.hasError("required"))
-      {
+      @if (submitted && form.get('userName')?.hasError("required")) {
       <span class="text-danger"> Username é um campo obrigatório. </span>
-      } @else if (submitted &&
-      registrationForm.get('userName')?.hasError("minlength") ||
-      registrationForm.get('userName')?.hasError("maxlength")) {
+      } @else if (submitted && form.get('userName')?.hasError("minlength") ||
+      form.get('userName')?.hasError("maxlength")) {
       <span class="text-danger">
         Username deve ter entre 5 e 50 caracteres.
       </span>
@@ -107,13 +103,9 @@
       <p-password
         formControlName="password"
         [toggleMask]="true"
-        [class.is-invalid]="
-          submitted && registrationForm.get('password')?.errors
-        "
+        [class.is-invalid]="submitted && form.get('password')?.errors"
         class="{{
-          submitted && registrationForm.get('password')?.errors
-            ? 'ng-invalid ng-dirty'
-            : ''
+          submitted && form.get('password')?.errors ? 'ng-invalid ng-dirty' : ''
         }}"
         styleClass="w-100"
       >
@@ -133,12 +125,10 @@
         </ng-template></p-password
       >
 
-      @if (submitted && registrationForm.get('password')?.hasError("required"))
-      {
+      @if (submitted && form.get('password')?.hasError("required")) {
       <span class="text-danger"> Password é um campo obrigatório. </span>
-      } @else if (submitted &&
-      registrationForm.get('password')?.hasError("minlength") ||
-      registrationForm.get('password')?.hasError("maxlength")) {
+      } @else if (submitted && form.get('password')?.hasError("minlength") ||
+      form.get('password')?.hasError("maxlength")) {
       <span class="text-danger">
         Password deve ter entre 5 e 50 caracteres.
       </span>
@@ -155,12 +145,10 @@
         autocomplete="off"
         formControlName="confirmPassword"
         [toggleMask]="true"
-        [class.is-invalid]="
-          submitted && registrationForm.get('confirmPassword')?.errors
-        "
+        [class.is-invalid]="submitted && form.get('confirmPassword')?.errors"
         styleClass="w-100"
         class=" {{
-          submitted && registrationForm.get('confirmPassword')?.errors
+          submitted && form.get('confirmPassword')?.errors
             ? 'ng-invalid ng-dirty'
             : ''
         }}"
@@ -178,12 +166,11 @@
           </ul>
         </ng-template></p-password
       >
-      @if (submitted &&
-      registrationForm.get('confirmPassword')?.hasError("required")) {
+      @if (submitted && form.get('confirmPassword')?.hasError("required")) {
       <span class="text-danger">
         Confirm Password é um campo obrigatório.
       </span>
-      } @if (submitted && registrationForm.errors?.['mismatchedPasswords']) {
+      } @if (submitted && form.errors?.['mismatchedPasswords']) {
       <span class="text-danger"> As palavras-passe não coincidem. </span>
       }
     </div>
@@ -201,7 +188,7 @@
     <button type="submit" class="btn-main" (click)="onSubmit()">
       @if (loading){
       <span class="spinner-border spinner-border-sm me-1"></span>
-      } Registar
+      } {{ isUpdate ? "Atualizar" : "Registar" }}
     </button>
   </div>
   <!-- end submit button -->
